rename `managed` to `autoVisible`
`managed` used to handle disposals as well. now this has to be handled manually. `autoVisible` only handles visibility. automatic disposals sometime run out of order and disposes handle created for a newly created state. this was observed while switching between 0 and double.infinity values of the container that holds webview.